fix: `tree` object is more memory efficient
The previous tree object used to be a vector of vectors structured, where the first vectors were the levels inside the tree and the second vectors the nodes stored inside that level.

The previous implementation was not memory efficient since it would create inner factors of dimension `n` which is the number of entries in the tree for all the levels.

The current implementation reduce the dimension of the inner vector to the amount of nodes that are actually stored in that level, which is identified by the variable `nodes_in_level`