fix: remove use of `parallelization_threshold` variable
The `parallelization_threshold` variable was added because I was afraid that building in parallel levels with a only a few hashing (the ones when you get close to the merkle root) would actually downgrade the performance of the library. But that's not the case so I removed this. Now all the levels inside a tree will be built in parallel